refactor: Update revenue field to BigIntegerField in Movie model and adjust related API services style: Modify IMDb 250 header for consistency and update navigation links in layout feat: Enhance search result view to include average rating and refactor API service calls

+++ CineValue/models.py
@@ -18,7 +18,7 @@ class Movie(models.Model):
     
     adult = models.BooleanField(default=False)
     budget = models.BigIntegerField(null=True, blank=True)
-    revenue = models.IntegerField(null=True, blank=True)
+    revenue = models.BigIntegerField(null=True, blank=True)
 
     genres = models.CharField(max_length=500, blank=True)
